+++ main/core/Resources/modules/plugin.js
@@ -95,7 +95,7 @@ registry.add('ClarolineCoreBundle', {
     'home'            :   () => { return import(/* webpackChunkName: "core-tool-home" */       '#/main/core/tools/home') },
     'workspaces'      :   () => { return import(/* webpackChunkName: "core-tool-workspaces" */ '#/main/core/tools/workspaces') },
     'resource_manager':   () => { return import(/* webpackChunkName: "core-tool-resources" */  '#/main/core/tools/resources') },
-    'parameters'      :   () => { return import(/* webpackChunkName: "core-tool-parameters" */  '#/main/core/tools/desktop-parameters') }
+    'parameters'      :   () => { return import(/* webpackChunkName: "core-tool-parameters" */  '#/main/core/tools/parameters') }
   },
 
   /**
