--- main/core/Resources/modules/tools/desktop-parameters/store/index.js
@@ -1,11 +1,0 @@
-/**
- * Desktop parameters store.
- */
-
-import {reducer} from '#/main/core/tools/desktop-parameters/store/reducer'
-import {selectors} from '#/main/core/tools/desktop-parameters/store/selectors'
-
-export {
-  reducer,
-  selectors
-}
