--- main/core/Resources/modules/tools/desktop-parameters/token/store/index.js
@@ -1,11 +1,0 @@
-/**
- * Desktop parameters store.
- */
-
-import {reducer} from '#/main/core/tools/desktop-parameters/token/store/reducer'
-import {selectors} from '#/main/core/tools/desktop-parameters/token/store/selectors'
-
-export {
-  reducer,
-  selectors
-}
