+++ main/core/Resources/modules/tools/parameters/store/index.js
@@ -1,0 +1,11 @@
+/**
+ * Desktop parameters store.
+ */
+
+import {reducer} from '#/main/core/tools/parameters/store/reducer'
+import {selectors} from '#/main/core/tools/parameters/store/selectors'
+
+export {
+  reducer,
+  selectors
+}
