+++ main/core/Resources/modules/tools/parameters/store/reducer.js
@@ -1,17 +1,16 @@
 import {makeInstanceAction} from '#/main/app/store/actions'
 import {makeFormReducer} from '#/main/app/content/form/store/reducer'
-import {reducer as tokenReducer} from '#/main/core/tools/desktop-parameters/token/store/reducer'
+import {reducer as tokenReducer} from '#/main/core/tools/parameters/token/store/reducer'
 import {combineReducers, makeReducer} from '#/main/app/store/reducer'
 
 import {TOOL_LOAD} from '#/main/core/tool/store/actions'
+import {selectors} from '#/main/core/tools/parameters/store'
 
 const reducer = combineReducers({
   tools: makeReducer([], {
-    [makeInstanceAction(TOOL_LOAD, 'parameters')]: (state, action) => action.toolData.tools
+    [makeInstanceAction(TOOL_LOAD, selectors.STORE_NAME)]: (state, action) => action.toolData.tools
   }),
-  editable: makeReducer(false, {
-  }),
-  toolsConfig: makeFormReducer('parameters.toolsConfig'),
+  toolsConfig: makeFormReducer(selectors.STORE_NAME+'.toolsConfig'),
   tokens: tokenReducer
 })
