+++ main/core/Resources/modules/tools/parameters/token/components/token.jsx
