+++ main/core/Resources/modules/tools/parameters/token/store/actions.js
