+++ main/core/Resources/modules/tools/parameters/token/store/index.js
@@ -1,0 +1,11 @@
+/**
+ * Desktop parameters store.
+ */
+
+import {reducer} from '#/main/core/tools/parameters/token/store/reducer'
+import {selectors} from '#/main/core/tools/parameters/token/store/selectors'
+
+export {
+  reducer,
+  selectors
+}
