+++ main/core/Resources/modules/tools/parameters/token/store/reducer.js
